upcoming: [DI-22596] - Alert detail chips
